clean(turret): Remove unnecessary functions
Control means that are told to targeet a specific position. If two systems ened to communicate the target position, the controlled system should not be used to communicate it. In a majority of situations, this is also not necessary. The need to obtain the target position or speed is simple not a thing in RMNA robotics programing.